- moved content track list out of InlineEditorProviders to InlineEditor.js - JSLintCleanup - pushed more inline editor code out of CSSInlineEditor - made CSSInlineEditor provider register itself without calling "init"
