+++ src/editor/CSSInlineEditor.js
@@ -10,36 +10,11 @@ define(function (require, exports, module) {
 
     // Load dependent modules
     var DocumentManager     = require("document/DocumentManager"),
+        HTMLUtils           = require("language/HTMLUtils"),
+        CSSUtils            = require("language/CSSUtils"),
     	EditorManager       = require("editor/EditorManager"),
     	InlineEditor 		= require("editor/InlineEditor");
 
-
-   /**
-     * Create the shadowing and filename tab for an inline editor.
-     * TODO (issue #424): move to createInlineEditorFromText()
-     * @private
-     */
-    function _createInlineEditorDecorations(editor, doc) {
-        // create the filename div
-        var filenameDiv = $('<div class="filename" style="visibility: hidden"/>')
-            .append('<div class="dirty-indicator"/>')
-            .append(doc.file.name);
-        
-        // add inline editor styling
-        // FIXME (jasonsj): #424, shadow only seems to work on scroller element and not on wrapper div
-        $(editor.getScrollerElement())
-            .append('<div class="shadow top"/>')
-            .append('<div class="shadow bottom"/>')
-            .append(filenameDiv);
-
-        // update the current inline editor immediately
-        // use setTimeout to allow filenameDiv to render first
-        setTimeout(function () {
-            _updateInlineEditorFilename(_editorHolderWidth(), filenameDiv);
-            _showDirtyIndicator(filenameDiv.find(".dirty-indicator"), doc.isDirty);
-            filenameDiv.css("visibility", "");
-        }, 0);
-    }
 
     /**
      * @constructor
@@ -52,8 +27,9 @@ define(function (require, exports, module) {
     CSSInlineEditor.prototype.constructor = CSSInlineEditor;
     CSSInlineEditor.prototype.parentClass = InlineEditor.InlineEditor;
     
-    // TY TODO: rename load?
+
     /** @param {!Editor} hostEditor  Outer Editor instance that inline editor will sit within.
+     * TY TODO: factor out css specific code from inline editor code
     */
     CSSInlineEditor.prototype.load = function (hostEditor) {
         this.hostEditor = hostEditor;
@@ -89,32 +65,20 @@ define(function (require, exports, module) {
         // wrapper div for inline editor
         this.htmlContent = htmlContent;
 
-        // TY TODO: this probably belongs on Editor, but it needs to assigned after this.editor is valid
+        // TY TODO: this probably belongs in InlineEditor, but it needs to assigned after this.editor is valid
         // When text is edited, auto-resize UI and sync changes to a backing full-size editor
         $(this.editor).on("change", function () {
             self.sizeInlineEditorToContents();
         });
         
         // TODO (jasonsj): XD
-        // TODO TY: where should this go?
-        _createInlineEditorDecorations(inlineInfo.editor, rule.document);
+        // TODO TY: review how this works
+        this.createInlineEditorDecorations(inlineInfo.editor, rule.document);
         
-        _htmlToCSSProviderContent.push(inlineInfo.content);
-        
-        // Manually position filename div's. Can't use CSS positioning in this case
-        // since the label is relative to the window boundary, not CodeMirror.
-        if (_htmlToCSSProviderContent.length > 0) {
-            $(window).on("resize", _updateAllFilenames);
-            $(DocumentManager).on("dirtyFlagChange", _dirtyFlagChangeHandler);
-        }
+        InlineEditor.addInlineEditorContent(inlineInfo.content);
         
         return (new $.Deferred()).resolve();
     };
-
-    CSSInlineEditor.prototype.onClosed = function () {
-        this.parentClass.prototype.onClosed.call(this); // call super.onClosed()
-        _inlineEditorRemoved(this.content);
-    }
     
     CSSInlineEditor.prototype.getRules = function () {
     };
@@ -132,5 +96,112 @@ define(function (require, exports, module) {
     };
 
     exports.CSSInlineEditor = CSSInlineEditor;
+
+
+    /**
+     * Given a position in an HTML editor, returns the relevant selector for the attribute/tag
+     * surrounding that position, or "" if none is found.
+     * @param {!Editor} editor
+     * @private
+     */
+    function _getSelectorName(editor, pos) {
+        var tagInfo = HTMLUtils.getTagInfo(editor, pos),
+            selectorName = "";
+        
+        if (tagInfo.position.tokenType === HTMLUtils.TAG_NAME) {
+            // Type selector
+            selectorName = tagInfo.tagName;
+        } else if (tagInfo.position.tokenType === HTMLUtils.ATTR_NAME ||
+                   tagInfo.position.tokenType === HTMLUtils.ATTR_VALUE) {
+            if (tagInfo.attr.name === "class") {
+                // Class selector. We only look for the class name
+                // that includes the insertion point. For example, if
+                // the attribute is: 
+                //   class="error-dialog modal hide"
+                // and the insertion point is inside "modal", we want ".modal"
+                var attributeValue = tagInfo.attr.value;
+                var startIndex = attributeValue.substr(0, tagInfo.position.offset).lastIndexOf(" ");
+                var endIndex = attributeValue.indexOf(" ", tagInfo.position.offset);
+                selectorName = "." +
+                    attributeValue.substring(
+                        startIndex === -1 ? 0 : startIndex + 1,
+                        endIndex === -1 ? attributeValue.length : endIndex
+                    );
+                
+                // If the insertion point is surrounded by space, selectorName is "."
+                // Check for that here
+                if (selectorName === ".") {
+                    selectorName = "";
+                }
+            } else if (tagInfo.attr.name === "id") {
+                // ID selector
+                selectorName = "#" + tagInfo.attr.value;
+            }
+        }
+        
+        return selectorName;
+    }
+
+    /**
+     * When cursor is on an HTML tag name, class attribute, or id attribute, find associated
+     * CSS rules and show (one/all of them) in an inline editor.
+     *
+     * @param {!Editor} editor
+     * @param {!{line:Number, ch:Number}} pos
+     * @return {$.Promise} a promise that will be resolved with:
+            TY TODO: update return comment to specify inline editor
+     *      {{content:DOMElement, height:Number, onAdded:function(inlineId:Number), onClosed:function()}}
+     *      or null if we're not going to provide anything.
+     */
+    function htmlToCSSProvider(hostEditor, pos) {
+        // Only provide a CSS editor when cursor is in HTML content
+        if (hostEditor._codeMirror.getOption("mode") !== "htmlmixed") {
+            return null;
+        }
+        var htmlmixedState = hostEditor._codeMirror.getTokenAt(pos).state;
+        if (htmlmixedState.mode !== "html") {
+            return null;
+        }
+        
+        // Only provide CSS editor if the selection is an insertion point
+        var sel = hostEditor.getSelection(false);
+        if (sel.start.line !== sel.end.line || sel.start.ch !== sel.end.ch) {
+            return null;
+        }
+        
+        var selectorName = _getSelectorName(hostEditor, pos);
+        if (selectorName === "") {
+            return null;
+        }
+
+        var result = new $.Deferred();
+
+        CSSUtils.findMatchingRules(selectorName)
+            .done(function (rules) {
+                if (rules && rules.length > 0) {
+                    var cssInlineEditor = new CSSInlineEditor(rules);
+                    
+                    cssInlineEditor.load(hostEditor).done(function () {
+                        result.resolve(cssInlineEditor);
+                    }).fail(function () {
+                        result.reject();
+                    });
+                } else {
+                    // No matching rules were found.
+                    result.reject();
+                }
+            })
+            .fail(function () {
+                console.log("Error in findMatchingRules()");
+                result.reject();
+            });
+        
+        return result.promise();
+    }
+
+
+
+    EditorManager.registerInlineEditProvider(htmlToCSSProvider);
+    
 
 });
