+++ src/editor/InlineEditorProviders.js
@@ -12,232 +12,10 @@
 define(function (require, exports, module) {
     'use strict';
     
-    // Load dependent modules
-    var HTMLUtils           = require("language/HTMLUtils"),
-        CSSUtils            = require("language/CSSUtils"),
-        DocumentManager     = require("document/DocumentManager"),
-        EditorManager       = require("editor/EditorManager"),
-        FileUtils           = require("file/FileUtils"),
-        CSSInlineEditor     = require("editor/CSSInlineEditor"),
-        ProjectManager      = require("project/ProjectManager");
-
-    // track divs to re-position manually
-    var _htmlToCSSProviderContent   = [];
-    
-    /**
-     * Reposition a .inlineCodeEditor .filename div { right: 20px; }
-     * @private
-     */
-    function _updateInlineEditorFilename(holderWidth, filenameDiv) {
-        var filenameWidth = $(filenameDiv).outerWidth();
-        $(filenameDiv).css("left", holderWidth - filenameWidth - 40);
-    }
-    
-    /**
-     * Shows or hides the dirty indicator
-     * @private
-     */
-    function _showDirtyIndicator($indicatorDiv, isDirty) {
-        // Show or hide the dirty indicator by adjusting
-        // the width of the div. The "hidden" width is 
-        // 4 pixels to make the padding look correct.
-        $indicatorDiv.css("width", isDirty ? 16 : 4);
-    }
-    
-    /**
-     * Returns editor holder width (not CodeMirror's width).
-     * @private
-     */
-    function _editorHolderWidth() {
-        return $("#editorHolder").width();
-    }
-    
-    /**
-     * Reposition all filename divs after a window resize.
-     * @private
-     */
-    function _updateAllFilenames() {
-        var holderWidth = _editorHolderWidth();
-        
-        _htmlToCSSProviderContent.forEach(function (value) {
-            var filenameDiv = $(value).find(".filename");
-            _updateInlineEditorFilename(holderWidth, filenameDiv);
-        });
-    }
-    
-    /**
-     * Respond to dirty flag change event. If the dirty flag is associated with an inline editor,
-     * show (or hide) the dirty indicator.
-     * @private
-     */
-    function _dirtyFlagChangeHandler(event, doc) {
-        
-        _htmlToCSSProviderContent.forEach(function (value) {
-            var $filenameDiv = $(value).find(".filename");
-            // This only checks filename here. If there are multiple documents with the same filename
-            // (in different directories), this test will fail.
-            // TODO: This needs to be fixed by connecting this method with a specific inline editor's state
-            if ($filenameDiv.text() === doc.file.name) {
-                _showDirtyIndicator($filenameDiv.find(".dirty-indicator"), doc.isDirty);
-            }
-        });
-    }
-    
-    /**
-     * Stops tracking an editor after being removed from the document.
-     * @private
-     */
-    function _inlineEditorRemoved(inlineContent) {
-        var indexOf = _htmlToCSSProviderContent.indexOf(inlineContent);
-        
-        if (indexOf >= 0) {
-            _htmlToCSSProviderContent.splice(indexOf, 1);
-        }
-        
-        // stop listening for resize when all inline editors are closed
-        if (_htmlToCSSProviderContent.length === 0) {
-            $(window).off("resize", _updateAllFilenames);
-            $(DocumentManager).off("dirtyFlagChange", _dirtyFlagChangeHandler);
-        }
-    }
-
-    
-    /**
-     * Given a position in an HTML editor, returns the relevant selector for the attribute/tag
-     * surrounding that position, or "" if none is found.
-     * @param {!Editor} editor
-     * @private
-     */
-    function _getSelectorName(editor, pos) {
-        var tagInfo = HTMLUtils.getTagInfo(editor, pos),
-            selectorName = "";
-        
-        if (tagInfo.position.tokenType === HTMLUtils.TAG_NAME) {
-            // Type selector
-            selectorName = tagInfo.tagName;
-        } else if (tagInfo.position.tokenType === HTMLUtils.ATTR_NAME ||
-                   tagInfo.position.tokenType === HTMLUtils.ATTR_VALUE) {
-            if (tagInfo.attr.name === "class") {
-                // Class selector. We only look for the class name
-                // that includes the insertion point. For example, if
-                // the attribute is: 
-                //   class="error-dialog modal hide"
-                // and the insertion point is inside "modal", we want ".modal"
-                var attributeValue = tagInfo.attr.value;
-                var startIndex = attributeValue.substr(0, tagInfo.position.offset).lastIndexOf(" ");
-                var endIndex = attributeValue.indexOf(" ", tagInfo.position.offset);
-                selectorName = "." +
-                    attributeValue.substring(
-                        startIndex === -1 ? 0 : startIndex + 1,
-                        endIndex === -1 ? attributeValue.length : endIndex
-                    );
-                
-                // If the insertion point is surrounded by space, selectorName is "."
-                // Check for that here
-                if (selectorName === ".") {
-                    selectorName = "";
-                }
-            } else if (tagInfo.attr.name === "id") {
-                // ID selector
-                selectorName = "#" + tagInfo.attr.value;
-            }
-        }
-        
-        return selectorName;
-    }
-    
-    /**
-     * Create the shadowing and filename tab for an inline editor.
-     * TODO (issue #424): move to createInlineEditorFromText()
-     * @private
-     */
-    function _createInlineEditorDecorations(editor, doc) {
-        // create the filename div
-        var filenameDiv = $('<div class="filename" style="visibility: hidden"/>')
-            .append('<div class="dirty-indicator"/>')
-            .append(doc.file.name);
-        
-        // add inline editor styling
-        // FIXME (jasonsj): #424, shadow only seems to work on scroller element and not on wrapper div
-        $(editor.getScrollerElement())
-            .append('<div class="shadow top"/>')
-            .append('<div class="shadow bottom"/>')
-            .append(filenameDiv);
-
-        // update the current inline editor immediately
-        // use setTimeout to allow filenameDiv to render first
-        setTimeout(function () {
-            _updateInlineEditorFilename(_editorHolderWidth(), filenameDiv);
-            _showDirtyIndicator(filenameDiv.find(".dirty-indicator"), doc.isDirty);
-            filenameDiv.css("visibility", "");
-        }, 0);
-    }
-    
-        
-    
-    /**
-     * When cursor is on an HTML tag name, class attribute, or id attribute, find associated
-     * CSS rules and show (one/all of them) in an inline editor.
-     *
-     * @param {!Editor} editor
-     * @param {!{line:Number, ch:Number}} pos
-     * @return {$.Promise} a promise that will be resolved with:
-            TY TODO: update return comment to specify inline editor
-     *      {{content:DOMElement, height:Number, onAdded:function(inlineId:Number), onClosed:function()}}
-     *      or null if we're not going to provide anything.
-     */
-    function htmlToCSSProvider(hostEditor, pos) {
-        // Only provide a CSS editor when cursor is in HTML content
-        if (hostEditor._codeMirror.getOption("mode") !== "htmlmixed") {
-            return null;
-        }
-        var htmlmixedState = hostEditor._codeMirror.getTokenAt(pos).state;
-        if (htmlmixedState.mode !== "html") {
-            return null;
-        }
-        
-        // Only provide CSS editor if the selection is an insertion point
-        var sel = hostEditor.getSelection(false);
-        if (sel.start.line !== sel.end.line || sel.start.ch !== sel.end.ch) {
-            return null;
-        }
-        
-        var selectorName = _getSelectorName(hostEditor, pos);
-        if (selectorName === "") {
-            return null;
-        }
-
-        var result = new $.Deferred();
-
-        CSSUtils.findMatchingRules(selectorName)
-            .done(function (rules) {
-                if (rules && rules.length > 0) {
-                    var cssInlineEditor = new CSSInlineEditor.CSSInlineEditor(rules);
-                    
-                    cssInlineEditor.load(hostEditor).done(function () {
-                        result.resolve(cssInlineEditor);
-                    }).fail(function () {
-                        result.reject();
-                    });
-                } else {
-                    // No matching rules were found.
-                    result.reject();
-                }
-            })
-            .fail(function () {
-                console.log("Error in findMatchingRules()");
-                result.reject();
-            });
-        
-        return result.promise();
-    }
 
 
-    /** Initialize: register listeners */
-    function init() {
-        EditorManager.registerInlineEditProvider(htmlToCSSProvider);
-    }
     
-    // Define public API
-    exports.init = init;
+
+    
+    
 });
